# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rotaract South Asia Analytics API",
    "version": "1.1.0",
    "description": "Public, high-performance REST API and Model Context Protocol (MCP) server providing open analytics data for Rotaract and Rotary across South Asia (Zones 4, 5, 6, and 7; 44 districts; 2,870+ clubs).\n\n### Model Context Protocol (MCP) Integration\nThis service includes a Model Context Protocol (MCP) server for direct use by AI assistants (Cursor, Claude Desktop, Antigravity, VS Code):\n- **Local Stdio Command:** `npm run mcp` or `node scripts/mcp_server.js`\n- **Remote HTTP Endpoint:** `POST https://insights.rsmda.org/api/mcp` (JSON-RPC 2.0)\n- **Available AI Tools (14 tools):** `get_summary`, `get_leaderboards`, `get_districts`, `get_district_insights`, `get_zone_summary`, `search_clubs`, `get_club_profile`, `get_new_clubs`, `find_compliance_risks`, `find_dual_risk_clubs`, `get_interact_analytics`, `find_rotary_opportunities`, `get_foundation_giving`, `get_worldwide_rankings`.\n\nAll endpoints are open, unauthenticated, and CORS-enabled.",
    "contact": {
      "name": "Rotaract South Asia Analytics Team",
      "url": "https://insights.rsmda.org"
    },
    "license": {
      "name": "Open Data Commons Open Database License (ODbL)",
      "url": "https://opendatacommons.org/licenses/odbl/"
    }
  },
  "servers": [
    {
      "url": "https://insights.rsmda.org",
      "description": "Production server (Primary)"
    },
    {
...
    },
    {
...
    }
...
  "paths": {
...
    "/api/v1/districts": {
      "get": {
        "tags": [
          "Districts"
        ],
        "summary": "List All Districts",
        "description": "Returns all 44 districts with key metrics and leadership contacts.",
        "operationId": "getDistricts",
        "parameters": [
          {
            "name": "zone",
            "in": "query",
            "required": false,
            "description": "Optional zone filter (e.g. 'Zone 4' or '4')",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "description": "Sort field (totalClubs, members, arrearsClubs, outstanding, trf, newClubs, interactGrowth, district)",
            "schema": {
              "type": "string",
              "default": "district"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "description": "Sort order (asc, desc)",
            "schema": {
              "type": "string",
              "default": "asc"
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "description": "Optional country filter (e.g. \"India\", \"Nepal\", \"Sri Lanka\")",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of districts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "integer"
                    },
                    "districts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DistrictSummary"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-openai-isConsequential": false,
        "x-ms-require-user-confirmation": false
      }
    },
...
    "/api/v1/clubs": {
      "get": {
        "tags": [
          "Clubs"
        ],
        "summary": "Search & Filter Clubs",
        "description": "Full-text search and multi-facet filtering across 2,870+ Rotaract clubs in South Asia.",
        "operationId": "searchClubs",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "Search keyword matching club name, club ID, sponsor club, or country",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "district",
            "in": "query",
            "required": false,
            "description": "Filter by district number (e.g. '3000')",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "zone",
            "in": "query",
            "required": false,
            "description": "Filter by RI zone (e.g. 'Zone 5')",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "base",
            "in": "query",
            "required": false,
            "description": "Filter by base type ('Community' or 'University')",
            "schema": {
              "type": "string",
              "enum": [
                "Community",
                "University"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Filter by club status ('Active' or 'Terminated')",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isArrears",
            "in": "query",
            "required": false,
            "description": "Filter by financial arrears status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isAtRisk",
            "in": "query",
            "required": false,
            "description": "Filter by high termination risk status (>= $75 USD outstanding)",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isNoOfficers",
            "in": "query",
            "required": false,
            "description": "Filter by missing reported club officers status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isNewClub",
            "in": "query",
            "required": false,
            "description": "Filter newly chartered clubs",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "description": "Sort field ('members', 'outstanding', 'name', 'id', 'charterDate', 'district')",
            "schema": {
              "type": "string",
              "enum": [
                "members",
                "outstanding",
                "name",
                "id",
                "charterDate",
                "district"
              ]
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "description": "Sort order ('desc' for largest/highest first, 'asc' for lowest/alphabetical)",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "minMembers",
            "in": "query",
            "required": false,
            "description": "Filter clubs with at least this many members",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "maxMembers",
            "in": "query",
            "required": false,
            "description": "Filter clubs with at most this many members",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "minOutstanding",
            "in": "query",
            "required": false,
            "description": "Filter clubs with outstanding dues in INR at least this amount",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of records to return (1-100, default 25)",
            "schema": {
              "type": "integer",
              "default": 25
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "Record offset for pagination (default 0)",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "maxOutstanding",
            "in": "query",
            "required": false,
            "description": "Maximum outstanding dues in INR",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of clubs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubListResponse"
                }
              }
            }
          }
        },
        "x-openai-isConsequential": false,
        "x-ms-require-user-confirmation": false
      }
    },
...
    "/api/v1/interact": {
      "get": {
        "tags": [
          "Overview"
        ],
        "summary": "Interact Analytics & Rotaract Sponsorships",
        "description": "Returns comprehensive Interact club analytics across South Asia: 8,921 total clubs, 735 suspended, 137 clubs sponsored by 65 Rotaract clubs, plus district/zone growth trends.",
        "operationId": "getInteractAnalytics",
        "parameters": [
          {
            "name": "district",
            "in": "query",
            "required": false,
            "description": "Optional district filter (e.g. 3000)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "zone",
            "in": "query",
            "required": false,
            "description": "Optional zone filter (e.g. Zone 5)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "description": "Optional country filter (e.g. \"India\", \"Nepal\", \"Sri Lanka\")",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "base",
            "in": "query",
            "required": false,
            "description": "Optional sponsoring Rotaract club base filter (\"Community\" or \"University\")",
            "schema": {
              "type": "string",
              "enum": [
                "Community",
                "University"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Interact analytics successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractAnalyticsResponse"
                }
              }
            }
          }
        },
        "x-openai-isConsequential": false,
        "x-ms-require-user-confirmation": false
      }
    },
...
    "/api/v1/compliance/unified": {
      "get": {
        "tags": [
          "Compliance"
        ],
        "summary": "Unified Compliance Issues Roster",
        "description": "Unified compliance endpoint querying clubs across all issue categories: financial arrears, missing officer reports, at-risk termination status, and dual-risk non-compliance.",
        "operationId": "getUnifiedComplianceIssues",
        "x-openai-isConsequential": false,
        "x-ms-require-user-confirmation": false,
        "parameters": [
          {
            "name": "issueType",
            "in": "query",
            "required": false,
            "description": "Filter by issue category: \"all\" (default), \"arrears\", \"missing_officers\", \"at_risk\", or \"dual_risk\"",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "arrears",
                "missing_officers",
                "at_risk",
                "dual_risk"
              ],
              "default": "all"
            }
          },
          {
            "name": "district",
            "in": "query",
            "required": false,
            "description": "Optional district number filter (e.g. \"3000\")",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "zone",
            "in": "query",
            "required": false,
            "description": "Optional zone filter (e.g. \"5\" or \"Zone 5\")",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "description": "Optional country filter (e.g. \"India\", \"Nepal\", \"Sri Lanka\")",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "base",
            "in": "query",
            "required": false,
            "description": "Optional club base filter (\"Community\" or \"University\")",
            "schema": {
              "type": "string",
              "enum": [
                "Community",
                "University"
              ]
            }
          },
          {
            "name": "minOutstanding",
            "in": "query",
            "required": false,
            "description": "Minimum outstanding dues in INR",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxOutstanding",
            "in": "query",
            "required": false,
            "description": "Maximum outstanding dues in INR",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "description": "Sort field (\"outstanding\", \"name\", \"district\")",
            "schema": {
              "type": "string",
              "enum": [
                "outstanding",
                "name",
                "district"
              ],
              "default": "outstanding"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "description": "Sort order (\"asc\" or \"desc\")",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Number of results to return (1-100, default 25)",
            "schema": {
              "type": "integer",
              "default": 25
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "Number of results to skip (default 0)",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unified compliance issues retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedComplianceResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "Total matching records"
          },
          "limit": {
            "type": "integer",
            "description": "Records per page limit"
          },
          "offset": {
            "type": "integer",
            "description": "Record offset index"
          },
          "hasMore": {
            "type": "boolean",
            "description": "True if more records exist"
          }
        },
        "required": [
          "total",
          "limit",
          "offset",
          "hasMore"
        ]
      },
...
      "DistrictSummary": {
        "type": "object",
        "properties": {
          "district": {
            "type": "string",
            "example": "3000"
          },
          "zone": {
            "type": "string",
            "example": "Zone 5"
          },
          "stats": {
            "type": "object"
          },
          "leadership": {
            "type": "object",
            "nullable": true,
            "properties": {
              "dg": {
                "type": "string"
              },
              "drr": {
                "type": "string"
              },
              "drc": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ClubSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "8824847"
          },
          "name": {
            "type": "string",
            "example": "Aarupadai Veedu Medical College & Hospital"
          },
          "district": {
            "type": "string",
            "example": "2981"
          },
          "zone": {
            "type": "string",
            "example": "Zone 5"
          },
          "base": {
            "type": "string",
            "example": "University"
          },
          "status": {
            "type": "string",
            "example": "Active"
          },
          "country": {
            "type": "string",
            "example": "India"
          },
          "sponsorClubs": {
            "type": "string",
            "example": "Pondicherry Beach Town (Rotary)"
          },
          "members": {
            "type": "integer",
            "example": 1
          },
          "isArrears": {
            "type": "boolean",
            "example": true
          },
          "outstandingINR": {
            "type": "number",
            "example": 2018
          },
          "outstandingUSD": {
            "type": "number",
            "example": 21.02
          },
          "isAtRisk": {
            "type": "boolean",
            "example": false
          },
          "isNoOfficers": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "ClubListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClubSummary"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        }
      },
...
      "InteractAnalyticsResponse": {
        "type": "object",
        "properties": {
          "scope": {
            "type": "object"
          },
          "overview": {
            "type": "object",
            "properties": {
              "totalInteractClubs": {
                "type": "integer",
                "example": 8921
              },
              "suspendedInteractClubs": {
                "type": "integer",
                "example": 735
              },
              "activeInteractClubs": {
                "type": "integer",
                "example": 8186
              },
              "rotaractClubsSponsoringInteractCount": {
                "type": "integer",
                "example": 65
              },
              "totalInteractClubsSponsoredByRotaract": {
                "type": "integer",
                "example": 137
              }
            }
          },
          "rotaractSponsors": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "districts": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "UnifiedComplianceResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      }
    }
  }
}